renamed lattice elements and flipped orderings. moved constants interfaces into their own module constants.rs. missing const interfaces now log info rather than panicking.
