+++ tests/opts/main.rs
@@ -3,6 +3,6 @@
 // We use pliron-llvm in this test, which is not supported in wasm.
 #![cfg(not(target_family = "wasm"))]
 
-mod constants;
 mod dce;
 mod mem2reg;
+mod sccp;
